requirements.txt
fastapi
uvicorn
websockets
pymongo
google-generativeai
pydantic
